Fixed bug:
When you create an anchor in the document for internal links, an extra paragraph was
being created.

Defined new "Reference" flowable that uses no space and use that instead.

Added a test case: before this revision, there was too much space between the two paragraphs (because of an empty paragraph in the middle).