Fixed some spacing issues, but there's a hard one left:
The frame besides the sidebar always starts with a new paragraph. However, since that paragraph is at "top-of-frame", it has no spaceBefore.

I can't fix it by adding topPadding to the frame because I don t know how much there should be.

I may have to make a flag to disable eating that space.