Possible fix for issues 261 and 264 checked in.
On the testing I do (without a lot of the add-ons installed)
the only changed output was on the files for these issues
(which now work) and also, for some reason, a minor change in
quickref.pdf.  Once the automated tester picks this up,
there might be some other minor changes as well.

Roberto, please verify that fix is acceptable.