The great style fix of 2019
I really don't like applying style checkers to existing code bases,
because they tend to make a complete mess of our ability to use
'git-blame' and related tools effectively. With that said, the code base
is in pretty bad shape owing to its sheer age and we'd otherwise be
fixing this stuff for years. Bite the bullet and run all the code
through three tools - black, autoflake8 and isort - meaning we don't
have to worry about this going forward.

    isort -rc .
    autoflake -r --in-place --remove-unused-variables --expand-star-imports .
    black -S .

pre-commit is integrated to ensure this can happen automatically going
forward.

Signed-off-by: Stephen Finucane <stephen@that.guru>