Change type of pdf_real_footnotes and pdf_use_toc options from str to bool
In case of pdf_use_toc, string "'True'" already evaluated to True.
However in the case of pdf_real_footnotes, "'False'" also evaluated
to True; which seems unintentional. So we now change the default to
actually be False.

This however means that we need to add the previous default to all
existing sphinx tests to prevent them from failing.