Add pytest-based test infrastructure
This is inspired by pytest's "Working with non-python tests" guide [1].
In short, we iterate through the 'input' directory and attempt to find
either files that end in '.txt' (meaning a test that should be run
through the 'rst2pdf' binary) or directories that start with 'sphinx-'
and contain a 'conf.py' file (meaning a test that should be run through
the 'sphinx-build' binary). In both cases, we build the PDFs and compare
checksums, just as we were doing before.

We *haven't* migrated a lot of the tooling to add additional checksums,
but that's okay since we can still do this manually and we plan to
switch to image-based comparisons in the near future.

[1] https://docs.pytest.org/en/5.4.1/example/nonpython.html

Signed-off-by: Stephen Finucane <stephen@that.guru>
Suggested-by: Norman Lorrain <normanlorrain@gmail.com>