Start using pymupdf to compare PDFs
Previously, we were assessing whether a PDF was correct by comparing a
checksum of the file against a list of known valid checksums. When a
different checksum was seen, which was quite often given the amount of
moving pieces here, it was up to the user to determine whether there was
a bug or simply a small but valid change in the output PDF. Let's
simplify this. Instead of storing checksums, use the pymupdf library to
compare the PDFs programmatically. We need to use a bit of fuzzing here,
since blocks can get moved around and text rewrapped, but it should be
mostly bullet proof otherwise.

Signed-off-by: Stephen Finucane <stephen@that.guru>