Run code through black
One should generally avoid applying large, style-only changes to
existing code bases because they tend to make a complete mess of our
ability to use 'git-blame' and related tools effectively. With that
said, the code base is in pretty bad shape owing to its sheer age and
we'd otherwise be fixing this stuff manually for years. Bite the bullet
and run all the code through black [1] (specifically, 'black -S', to
avoid unnecessary conversion of single quotes to double quotes) to
address the worst of the issues. A future change will address the
remaining issues like use of wildcard imports and commented out (dead)
code.

Note that black defaults to a max line length of 88, not 79 [2], which
means we need to configure flake8 to allow this.

[1] https://pypi.org/project/black/
[2] https://black.readthedocs.io/en/stable/the_black_code_style.html

Signed-off-by: Stephen Finucane <stephen@that.guru>
Closes: #724
Closes: #711