Add 'package' roles
We want to be able to incorporate the rst2pdf version information in the
manual automatically, without reinventing SVN's keyword substitution
feature (which is what docutils uses to do this) in Git. Add a pair of
new roles to do this for us. Because these are broadly useful, they are
included as part of rst2pdf itself rather than just our local docs.
They rely on the 'packaging' and 'importlib_metadata' packages, the
former of which is provided in a typical installation using 'setuptools'
(but is listed in our own requirements.txt just in case) and the latter
of which is already a requirement.

An unnecessary call to 'register_canonical_role' is removed from
'rst2html-manual.py', since the mere importing of the module has this
side-effect (for better or worse).

Signed-off-by: Stephen Finucane <stephen@that.guru>