image: Handle missing 'svglib'
This is apparently an optional import, but attempting to build a
document containing an SVG will cause things to explode in a long
stacktrace ending with:

  NameError: name 'SVGImage' is not defined

Resolve this failing cleanly with a user friendly warning. While we're
here, fix up some logs to use delayed interpolation, along with many
docstrings.

Signed-off-by: Stephen Finucane <stephen@that.guru>