Restore today setting to Sphinx cover page tests
The cover page on a Sphinx test have a date on it that defaults to
today's date. However, for the tests to pass we need the date to match the
date on the reference PDF. This is controlled by the Sphinx setting
`today` which we need to set.

This setting was inadvertently removed for this set of tests in PR #855
(commit 89c630b356d5bc40524c30b171629e6fe86e0885).