Ensure fontList is set for a previously embedded font
If a font is aliased and then that font name is subsequently used as a
style name, then the `fontList` variable will not be set which will
result in the font name becoming an alias to whichever font happened to
be in `fontList` the previous time around the loop.

Example style sheet:

```
fontsAlias:
    stdFont: inconsolata
    stdBold: carlito

styles:
    normal:
        fontName: inconsolata
        textColor: #383745
        fontSize: 16
        alignment: TA_LEFT

    heading:
        fontName: stdFont
        fontSize: 32
        textColor: #1a8099
```

In this example, `inconsolata` is set as an alias to `stdFont` which
means that `inconsolata` is registed as the key name for the embedded font list pointing to the correct
font file names for that font
(`Inconsolata-Regular_9ba2497f55c3435c544f6702d57e7a94` on my machine).
However, `inconsolata` is also referenced as the `fontName` for the
`normal` style which means that without this fix, the embedded font
referenced by the `inconsolata` key now points the last font that was
processed (`Cardo` in this case).

This was reported in issue [#902][1]

[1]: https://github.com/rst2pdf/rst2pdf/issues/902.