Fix #820: Use right anchor name for target nodes
A reference with another reference as its embedded URI yields a target
with either 'refid' or 'refuri' set, depending on whether it points to a
section or an URL, respectively.

Previously it was always assumed to have 'refuri', causing rst2pdf to
crash when the embedded reference was a section. To fix the issue, use
either 'refid' or 'refuri' as the anchor 'name' field, depending on
which is present in the target node.