Test that writing to stdout doesn't fail
When writing to stdout in a test, the PDF file is send to the log file,
but we do not yet have a way to test for content in that log file.
However we can test that the return code is zero as if the PDF is not
correctly sent to stdout, a non-zero status code will be returned by
rst2pdf.