Handle pygments styles in Sphinx and suppress warnings
Sphinx 4 uses pygments style names without the 'pygments' prefix that
rst2pdf uses, so if we fail to find a style name, check again with the
'pygments-' prefix and if we find that, use it.

Also introduce `suppress_undefined_style_warning` to Styles as
`HandleSphinxDesc` will try to find a style named after the Sphinx
domain's desctype. There are a lot of these as we don't need them to be
defined as the "normal" style is correct for them. As such, we suppress
the warning in this specific case as it's irrelevant.