Resize a table image if it is taller than the page
An image cannot be rendered across a page boundary and so if it is too
tall, then it needs to be resized. This happens automatically for plain
images, but not for those in a table (including lists and figures in
rst2pdf).

This is because the availHeight parameter passed to wrap() is set to a
very large value by reportlab on the assumption that a table cell's data
can be rendered over a page boundary. We therefore need to set the
correct maximum height into the MyImage instances within a DelayedTable
so that they can be resized correctly.