Warn if fc-match substitutes a different font
If the font name specified hasn't been embedded, then fc-match will look
for it. As it will always return something, warn the user if the
matched font name is not the same as what was asked for.

This is a warning as the font selected by fc-match may very well be good
enough for the user and so the PDF should be created.

To test for this, introduce a new `.expected_log` file that contains
strings to be searched for in the log output.