Write dependency file
Add new parameter --record-dependencies which works the same way was for
vanilla docutils rst2foo commands.

The target file will contain a list of files that were used to generate
the output file. The list can be consumed by build-systems to
efficiently regenerate the PDF if any of the inputs has changed.

Fixes #1108