Update tests for issue 255
It wasn't clear that the test for a table was still failing due to the
test copy not forcing a page break with later ReportLab versions.

Update the test for tables to show that it doesn't work and also create
a test for paragraphs that show that it works correctly in this case.