Set middleware.jpg's DPI to 300 to fix test_jpg_scaling
Prior to Pillow 11.1.0, the dpi for an image wasn't being populated when
using cm, so rst2pdf defaulted to 300, even though middleware.jpg was
set to 72dpi.

Hence, with Pillow 11.1.0+, test_jpg_scaling test fails as the image is
rendered at a different size. To ensure that the test continues to pass,
we set the dpi of the image to match the default that we were using.