Fix Sphinx productionlist handling
With newer Sphinx versions the formatting of a productionlist is now
done for us, so we use that. This has shown up that our previous
formatting had a double space after `::=` as the new formatting only has
a single one.

Also, in #1021, we changed the default code style, so I've added the
legacy style to this test to keep it consistent.