Copy docutils' utils/_roman_numerals.py to ourselves
With docutils 0.22, the original utils/roman.py was replaced with
utils/_roman_numerals.py. As this new file is prefixed with an
underscore, it's now internal to docutils and its API cannot be relied
upon.

As utils/_roman_numerals.py is licensed as Zero Clause BSD or public
domain, we have copied to roman_numerals.py so that we have a stable API
to use in rst2pdf with no new required dependencies.